chore: extract getProofOfDelegatedIdentity to wallet-core
# Conflicts:
#	suite-common/suite-sync/package.json
#	yarn.lock
