+++ suite-common/suite-sync/package.json
@@ -12,8 +12,6 @@
     },
     "dependencies": {
         "@evolu/common": "^6.0.1-preview.32",
-        "@noble/curves": "^1.9.7",
-        "@noble/hashes": "^1.6.1",
         "@reduxjs/toolkit": "2.10.1",
         "@suite-common/metadata-types": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
