+++ suite-common/wallet-core/package.json
@@ -12,6 +12,8 @@
     },
     "dependencies": {
         "@mobily/ts-belt": "^3.13.1",
+        "@noble/curves": "^1.9.7",
+        "@noble/hashes": "^1.6.1",
         "@reduxjs/toolkit": "2.10.1",
         "@suite-common/bluetooth": "workspace:*",
         "@suite-common/device-authenticity": "workspace:*",
