+++ suite-common/wallet-core/src/index.ts
@@ -9,6 +9,8 @@ export * from './blockchain/blockchainMiddleware';
 export * from './blockchain/blockchainReducer';
 export * from './blockchain/blockchainSelectors';
 export * from './blockchain/blockchainThunks';
+export * from './device/delegatedIdentityKey/getProofOfDelegatedIdentity';
+export * from './device/delegatedIdentityKey/retrieveDelegatedIdentityKeyThunk';
 export * from './device/deviceActions';
 export * from './device/deviceConstants';
 export * from './device/deviceReducer';
@@ -17,7 +19,6 @@ export * from './device/deviceThunks';
 export * from './device/deviceUtils';
 export * from './device/preparePushNotificationMiddleware';
 export * from './device/publicKeyActions';
-export * from './device/retrieveDelegatedIdentityKeyThunk';
 export * from './device/usePinHook';
 export * from './discovery/discoveryActions';
 export * from './discovery/discoveryReducer';
